SLVS-1853 Move class into its own file.
