+++ src/Roslyn.Suppressions/Roslyn.Suppressions.UnitTests/InProcess/IssueConverterTests.cs
@@ -31,7 +31,7 @@ namespace SonarLint.VisualStudio.Roslyn.Suppressions.UnitTests.InProcess
         {
             var sonarIssue = CreateSonarQubeIssue(hash: "aaaa", filePath: "\\bbb\\ccc\\file.txt", issueKey: "key");
 
-            var actual = RoslynSettingsFileSynchronizer.IssueConverter.Convert(sonarIssue);
+            var actual = IssueConverter.Convert(sonarIssue);
 
             actual.Hash.Should().Be("aaaa");
             actual.FilePath.Should().Be("\\bbb\\ccc\\file.txt");
@@ -47,7 +47,7 @@ namespace SonarLint.VisualStudio.Roslyn.Suppressions.UnitTests.InProcess
             // 1-based Sonar line numbers should be converted to 0-based Roslyn line numbers
             var sonarIssue = CreateSonarQubeIssue(line: sonarLineNumber);
 
-            var actual = RoslynSettingsFileSynchronizer.IssueConverter.Convert(sonarIssue);
+            var actual = IssueConverter.Convert(sonarIssue);
 
             if (expected.HasValue)
             {
@@ -79,7 +79,7 @@ namespace SonarLint.VisualStudio.Roslyn.Suppressions.UnitTests.InProcess
         {
             var sonarIssue = CreateSonarQubeIssue(ruleId: compositeKey);
 
-            var actual = RoslynSettingsFileSynchronizer.IssueConverter.Convert(sonarIssue);
+            var actual = IssueConverter.Convert(sonarIssue);
 
             actual.RoslynLanguage.Should().Be(expectedLanguage);
             actual.RoslynRuleId.Should().Be(expectedRuleKey);
