+++ src/Roslyn.Suppressions/Roslyn.Suppressions/InProcess/RoslynSettingsFileSynchronizer.cs
@@ -218,57 +218,4 @@ internal sealed class RoslynSettingsFileSynchronizer : IRoslynSettingsFileSynchr
             CodeMarkers.Instance.FileSynchronizerUpdateStop();
         }
     }
-
-    // Converts SonarQube issues to SuppressedIssues that can be compared more easily with Roslyn issues
-    internal static class IssueConverter
-    {
-        public static SuppressedIssue Convert(SonarQubeIssue issue)
-        {
-            var (repoKey, ruleKey) = GetRepoAndRuleKey(issue.RuleId);
-            var language = GetRoslynLanguage(repoKey);
-
-            var line = issue.TextRange == null ? (int?)null : issue.TextRange.StartLine - 1;
-            return new SuppressedIssue
-            {
-                RoslynRuleId = ruleKey,
-                FilePath = issue.FilePath,
-                Hash = issue.Hash,
-                RoslynLanguage = language,
-                RoslynIssueLine = line,
-                IssueServerKey = issue.IssueKey
-            };
-        }
-
-        private static (string repoKey, string ruleKey) GetRepoAndRuleKey(string sonarRuleId)
-        {
-            // Sonar rule ids are in the form "[repo key]:[rule key]"
-            var separatorPos = sonarRuleId.IndexOf(":", StringComparison.OrdinalIgnoreCase);
-            if (separatorPos > -1)
-            {
-                var repoKey = sonarRuleId.Substring(0, separatorPos);
-                var ruleKey = sonarRuleId.Substring(separatorPos + 1);
-
-                return (repoKey, ruleKey);
-            }
-
-            return (null, null); // invalid rule key -> ignore
-        }
-
-        private static RoslynLanguage GetRoslynLanguage(string repoKey)
-        {
-            // Currently the only Sonar repos which contain Roslyn analysis rules are
-            // csharpsquid and vbnet. These include "normal" and "hotspot" rules.
-            // The taint rules are in a different repo, and the part that is implemented
-            // as a Roslyn analyzer won't raise issues anyway.
-            switch (repoKey)
-            {
-                case "csharpsquid": // i.e. the rules in SonarAnalyzer.CSharp
-                    return RoslynLanguage.CSharp;
-                case "vbnet": // i.e. SonarAnalyzer.VisualBasic
-                    return RoslynLanguage.VB;
-                default:
-                    return RoslynLanguage.Unknown;
-            }
-        }
-    }
 }
